+++ CHANGELOG.rst
@@ -2,6 +2,9 @@ Changelog
 =========
 
 
+v0.10.0 (2017-02-13)
+--------------------
+
 - Add ``zulu.parse_delta`` as alias for ``Delta.parse``.
 - Deprecate ``zulu.delta`` in favor of ``zulu.parse_delta``.
 - Allow first argument to ``Zulu()``, ``Zulu.parse()``, and ``zulu.parse()`` to be a ``dict`` containing keys corresponding to initialization parameters.